Quality of Life Updates for New Season (#10)
* Start QOL updates

* Add auto scrim cancel on away/home channel/category deletion

* Fixes to cancelling scrim automatically

* Add edge case for team chat being deleted, fixes to func

* Add automatic team deletion system

* Remove UNIQUE constraint on team name

* Remove unused links module

* Add docs for infractions settings

* Fix emoji in snipe command

* Add a note for mutual guilded members

* Bulk snipe command to be hybrid for the mods

* Black isort, separate image module into own directory

* Refactor image request command:
- Move to new DB table
- Allow users to make many requests in one command
- Allow the command to be not FURY guild specific

* Remove unused profanity detector

* Clean requirements

* Add image request NSFW classifications

* Add check for attachment content type

* Refactor attachment submission message

* Black isort

* Make typing test command public

* Add WPM metric to typing test

* Remove redundant docs

* Remove unused constants and files

* Update Team.from_channel to raise better Exc

* Refactor team deletion method to include reason parameter

* message_tracking -> flvs, update how cache loaders work

* Clean how member notifier works

* Refactor images schema and tables

* spelling issue

* Refactor team deletion method to include reason parameter

* Refactor _append_nsfw_classification method to handle errors and notify developer

* Refactor image approval process to include additional message and fix image display

* Refactor practice end message formatting

* Refactor PracticeView interaction check to prevent interaction with ended practices

* Cleanup TODO comments, many of them done

* Refactor leaderboard display format to include team names and points

* Refactor scrim confirmation message formatting

* Remove scrim reschedule functionality

* Refactor team member removal and synchronization

* Refactor channel overwrites for team synchronization

* Refactor channel overwrites for team synchronization

* Refactor notification channel query to include enable_no_dms_open flag

* Move events/notifier to new infractions module

* Move to new module for hot loading

* Refactor image module to use updated NudeDetector

* Finalize panel for attachment request setting management

* Refactor PracticeLeaderboardCog's load_leaderboards method to handle missing guilds and channels

* Start base impl for infraction counter

* Start working on cleanup command

* Start working on infractions management panel

* Refactor dm_notifications.py and moderation.py

* Black isort

* Add base impl for infractions panel, need to cache

* Remove unneeded type ignores

* Add infractions settings local cache storage

* Add commands to count and show recent infractions

- Added two new commands to the `Infractions` cog: `infractions_count` and `infractions_recent`. These commands allow users to count the number of infractions a member has and show the hyperlink to their most recent infraction, respectively.

- The `infractions_count` command checks if the guild has infractions settings enabled and returns the total number of infractions for the specified member.

- The `infractions_recent` command also checks if the guild has infractions settings enabled and returns the hyperlink to the most recent infraction for the specified member.

- These commands are part of the ongoing development of the infractions feature and will be useful for moderation purposes.

- Updated the `Infractions` cog to include the new commands and handle the corresponding interactions.

- Added necessary database queries and table definitions to support the new functionality.

- This commit addresses the following tasks:
  - Add `infractions_count` command
  - Add `infractions_recent` command
  - Update `Infractions` cog
  - Update database schema

* Refactor imports in panel.py

* Refactor moderation.py to support cleaning up messages from a user in a specific channel

* Black isort

* Refactor moderation.py to update the type of the cleanup context command

* QOL fixes to error handler

* Fixes to logging errors

* Refactor panel.py to import InfractionsSettings from settings module

* Finalize infraction settings panel

* Refactor Infractions class to inherit from DmNotifications and InfractionCounter

* Refactor logging in InfractionCounter

* Refactor InfractionsSettings to include a method for clearing infractions

* Refactor Infractions class to include a method for clearing infractions

* Refactor Infractions class to include a method for clearing all infractions

* Add game, black isort

* Refactor logging in InfractionCounter to include target information

* Push workflow py version to 3.10

* Ignore virtual environment folder in .gitignore

* Upgrade pip and setuptools in autoblack_pull_request.yml

* Upgrade pip and setuptools in autoblack_pull_request.yml

---------

+++ cogs/teams/scrims/persistent.py
@@ -102,17 +102,18 @@ class HomeConfirm(discord.ui.View):
                 embed.add_field(name='Scrim Date and Time', value=scrim.scheduled_for_formatted())
             else:
                 embed = self.bot.Embed(
-                    title='Scrim Scheduled!',
+                    title='Scrim Scheduled',
                     description=f'A scrim on {scrim.scheduled_for_formatted()} **against {scrim.away_team.display_name}** has '
                     'been fully scheduled.',
                 )
                 embed.add_field(
                     name='How do I Scrim?',
-                    value='10 minutes before the scrim is scheduled to begin, '
+                    value='When the scrim is scheduled to begin, '
                     'FuryBot will create a chat for both teams to communicate. In this chat, '
                     f'the home team, **{scrim.home_team.display_name}**, will create the private match for the away team, '
                     f'**{scrim.away_team.display_name}**, to join with. You decide how much and long you want to play. The scrim channel '
-                    'will automatically be deleted after 5 hours. Please note you should use your private team voice chat for communication.',
+                    'will automatically be deleted after 5 hours. This chat is simply for the two teams to communicate, you should use '
+                    'your private team voice chat for communication.',
                     inline=False,
                 )
 
@@ -121,10 +122,7 @@ class HomeConfirm(discord.ui.View):
                     and len(self.scrim.away_confirm_anyways_voter_ids) >= self.scrim.per_team // 2
                 ):
                     force_started_value = f'The scrim was force started by: {", ".join(m.mention for m in self.scrim.away_confirm_anyways_voters)}'
-                else:
-                    force_started_value = 'The scrim was not force started.'
-
-                embed.add_field(name='Scrim Force Started?', value=force_started_value, inline=False)
+                    embed.add_field(name='Scrim Force Started?', value=force_started_value, inline=False)
 
             embed.add_field(
                 name='Confirmed Teammates',
@@ -181,12 +179,19 @@ class HomeConfirm(discord.ui.View):
             return await interaction.response.edit_message(view=self, embed=self.embed)
 
         # We need to change the status and update the message,
-        # all requied members have voted
+        # all required members have voted
         await self.scrim.change_status(ScrimStatus.pending_away)
         await interaction.response.edit_message(view=None, embed=self.embed)
 
         # Now send to the other team
         channel = self.scrim.away_team.text_channel
+        if not channel:
+            # The channel was deleted, we need to cancel the scrim
+            await self.scrim.cancel(reason='The away channel has been deleted.')
+            return await interaction.response.send_message(
+                'The away channel has been deleted. Cancelling the scrim...', ephemeral=True
+            )
+
         view = AwayConfirm(self.scrim)
         message = await channel.send(
             embed=view.embed, view=view, content='@everyone', allowed_mentions=discord.AllowedMentions.all()
@@ -287,7 +292,7 @@ class AwayForceConfirm(discord.ui.View):
                 'Hey, you aren\'t on this team so you can\'t vote!', ephemeral=True
             )
 
-        # Check if they've voted, if they havent then we need to yell at them
+        # Check if they've voted, if they haven't then we need to yell at them
         if interaction.user.id not in self.scrim.away_confirm_anyways_voter_ids:
             return await interaction.response.send_message(
                 'Hey, you haven\'t voted to force confirm this scrim yet!', ephemeral=True
@@ -306,7 +311,7 @@ class AwayForceConfirm(discord.ui.View):
         await self.scrim.edit(away_confirm_anyways_voter_ids=self.scrim.away_confirm_anyways_voter_ids)
 
         if len(self.scrim.away_confirm_anyways_voter_ids) < self.required_to_confirm:
-            # We don't have enough members to force econfirm
+            # We don't have enough members to force confirm
             return
 
         # We have enough! Confirm the scrim.
@@ -314,11 +319,20 @@ class AwayForceConfirm(discord.ui.View):
 
         # Delete this message
         away_channel = self.scrim.away_team.text_channel
+        if away_channel is None:
+            # For whatever reason, this away channel has been deleted. We must cancel this scrim as the data
+            # has been invalidated.
+            return await self.scrim.cancel(reason='The away channel has been deleted.')
+
         message = await away_channel.fetch_message(cast(int, self.scrim.away_confirm_anyways_message_id))
         await message.delete()
 
         # Now update the home message
         home_channel = self.scrim.home_team.text_channel
+        if home_channel is None:
+            # Similar to before, if the home channel is also deleted, we must cancel the scrim.
+            return await self.scrim.cancel(reason='The home channel has been deleted.')
+
         message = await home_channel.fetch_message(self.scrim.home_message_id)
         view = HomeConfirm(self.scrim)
         await message.edit(view=None, embed=view.embed)
@@ -382,11 +396,12 @@ class AwayConfirm(discord.ui.View):
             )
             embed.add_field(
                 name='How do I Scrim?',
-                value='10 minutes before the scrim is scheduled to begin, '
-                'FuryBot will create a chat for bothho teams to communicate. In this chat, '
+                value='When the scrim is scheduled to begin, '
+                'FuryBot will create a chat for both teams to communicate. In this chat, '
                 f'the home team, **{scrim.home_team.display_name}**, will create the private match for the away team, '
                 f'**{scrim.away_team.display_name}**, to join with. You decide how much and long you want to play. The scrim channel '
-                'will automatically be deleted after 5 hours. Please note you should use your private team voice chat for communication.',
+                'will automatically be deleted after 5 hours. This chat is simply for the two teams to communicate, you should use '
+                'your private team voice chat for communication.',
                 inline=False,
             )
             embed.add_field(name='Confirmed Teammates', value=', '.join(m.mention for m in scrim.away_voters), inline=False)
@@ -399,10 +414,7 @@ class AwayConfirm(discord.ui.View):
                 force_started_value = (
                     f'The scrim was force started by: {", ".join(m.mention for m in self.scrim.away_confirm_anyways_voters)}'
                 )
-            else:
-                force_started_value = 'The scrim was not force started.'
-
-            embed.add_field(name='Scrim Force Started?', value=force_started_value, inline=False)
+                embed.add_field(name='Scrim Force Started?', value=force_started_value, inline=False)
 
             return embed
 
@@ -452,12 +464,19 @@ class AwayConfirm(discord.ui.View):
 
         # And update the other message from the home team's chat:
         home_message = await self.scrim.home_message()
+        if home_message is None:
+            # If the home message has been deleted, we must cancel the scrim
+            await interaction.response.send_message(
+                'The home message has been deleted. Cancelling the scrim...', ephemeral=True
+            )
+            return await self.scrim.cancel(reason='The home message has been deleted.')
+
         view = HomeConfirm(self.scrim)
         await home_message.edit(embed=view.embed)
 
     @discord.ui.button(label='Force Confirm', custom_id='force-confirm-confirm')
     @default_button_doc_string
-    async def force_confirn(self, interaction: discord.Interaction[FuryBot], button: ButtonType[Self]) -> None:
+    async def force_confirm(self, interaction: discord.Interaction[FuryBot], button: ButtonType[Self]) -> None:
         """Launches a vote to force confirm the scrim. This can be done if the team is unable to get a full team
         but still wants to scrim."""
         if self.scrim.per_team < 2:
@@ -466,6 +485,12 @@ class AwayConfirm(discord.ui.View):
             )
 
         away_text_channel = self.scrim.away_team.text_channel
+        if away_text_channel is None:
+            # This channel has been deleted, we must cancel the scrim
+            await interaction.response.send_message(
+                'The away channel has been deleted. Cancelling the scrim...', ephemeral=True
+            )
+            return await self.scrim.cancel(reason='The away channel has been deleted.')
 
         if self.scrim.away_confirm_anyways_message_id:
             url = f'https://discordapp.com/channels/{away_text_channel.guild.id}/{away_text_channel.id}/{self.scrim.away_confirm_anyways_message_id}'
@@ -496,10 +521,10 @@ class AwayConfirm(discord.ui.View):
         )
         await self.scrim.edit(away_confirm_anyways_message_id=message.id)
 
-    @discord.ui.button(label='Unconfirm', custom_id='away-confirm-unconfirm')
+    @discord.ui.button(label='Un-Confirm', custom_id='away-confirm-un-confirm')
     @default_button_doc_string
-    async def unconfirm(self, interaction: discord.Interaction[FuryBot], button: ButtonType[Self]) -> None:
-        """Allows a member to unconfirm their vote to scrim at that given time."""
+    async def un_confirm(self, interaction: discord.Interaction[FuryBot], button: ButtonType[Self]) -> None:
+        """Allows a member to un-confirm their vote to scrim at that given time."""
         try:
             await self.scrim.remove_vote(interaction.user.id, self.scrim.away_id)
         except ValueError:
